fix: update component names
