+++ src/components/edit/exchange-update/index.js
@@ -1,2 +1,2 @@
-export { EditExchange } from './edit-exchange.js'
+export { ExchangeForm } from './exchange-form.js'
 export { useFetchExchange } from './useFetchExchange.js'
