+++ src/components/edit/request-update/index.js
@@ -1,2 +1,2 @@
-export { RequestEdit } from './edit-request.js'
+export { RequestForm } from './request-form.js'
 export { requestsReducer } from './requests-reducer.js'
