+++ src/components/edit/request-update/request-form.module.css
@@ -1,27 +1,44 @@
-.subsectionField {
+.editArea {
+    width: 100%;
+    overflow-y: auto;
+    height: calc(100% - 56px);
+    background: var(--colors-greyt00);
+}
+
+.editFormArea {
+    background: white;
+    border: 1px solid var(--colors-grey300);
+    padding: var(--spacers-dp12);
+}
+
+.editUpdateTitle {
+    font-size: 26px;
+    font-weight: 300;
+    padding: 0;
+}
+
+.editContainer {
+    padding: 0 var(--spacers-dp16) 0 var(--spacers-dp16);
+}
+
+.bottomBar {
+    height: 56px;
+    width: 100%;
+    padding: var(--spacers-dp12);
+    box-shadow: 0px 0px 1px rgba(33, 41, 52, 0.1),
+        0px -4px 6px -1px rgba(33, 41, 52, 0.1),
+        0px -2px 4px -1px rgba(33, 41, 52, 0.06);
+}
+
+.hidden {
+    display: none;
+}
+
+.fullHeight {
+    height: 100%;
+}
+
+.errorBoxContainer {
     margin-block-end: var(--spacers-dp16);
-}
-
-.subsectionField1000 {
-    margin-block-end: var(--spacers-dp16);
-    max-width: 1000px;
-}
-
-.subsectionField600 {
-    margin-block-end: var(--spacers-dp16);
-    max-width: 750px;
-}
-
-.subsectionBlockEnd {
-    margin-block-end: var(--spacers-dp24);
-}
-
-.visualizationSubtext {
-    font-size: 12px;
-    color: var(--colors-grey600);
-    margin-inline-start: 26px;
-}
-
-.checkContainer {
-    margin-block-start: var(--spacers-dp16);
+    width: 1000px;
 }
